Improved ISBN Clean & Validation, added Check Digit Calculation
Integrated a function “cleanISBN($$isbn)”, to improve the cleaning of
an incorrectly entered ISBN, now all non-numeric characters are
stripped. The result should be always number, therefore it is easier to
do a first validity check. (Validation of ISBN-10 or ISBN-13 can be
obtained by a simple modulo (10 or 11 respectively) of the “pondered”
sum.

Added new function to calculate check digit (useful to translate 10->13)

https://en.wikipedia.org/wiki/International_Standard_Book_Number#Check_d
igits